% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-graph.R
\name{cost_graph}
\alias{cost_graph}
\title{Weighted network of travel costs}
\usage{
cost_graph(r)
}
\arguments{
\item{r}{SpatRaster of energy costs.}
}
\value{
igraph object, the weighted graph of travel costs
}
\description{
Takes a raster of the energy landscape and builds the weighted network
with vertices the cells and edges connecting the 8 adjacent cells,
weighted by their cost of travel.
}
