% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem}
\alias{dem}
\title{Digital Elevation Model from Tinitaly}
\format{
A matrix with 501 rows and 501 columns of the DEM at 100 x 100
  meters.
}
\usage{
dem
}
\description{
Digital Elevation Model from Tinitaly
}
\keyword{datasets}
