% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edgelist}
\alias{edgelist}
\title{Edgelist of weighted graph based on energy landscapes.}
\usage{
edgelist(x)
}
\arguments{
\item{x}{matrix of the cost of travel.}
}
\description{
Edgelist of weighted graph based on energy landscapes.
}
