% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pa}
\alias{pa}
\title{Protected Areas for the 'dem' dataset}
\format{
A matrix with the protected areas from WDPA for the 'dem' dataset.
}
\usage{
pa
}
\description{
Protected Areas for the 'dem' dataset
}
\keyword{datasets}
