% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-functions.R
\name{parthian_paths}
\alias{parthian_paths}
\title{Calculate least-cost paths}
\usage{
parthian_paths(g, r, locations)
}
\arguments{
\item{g}{igraph object, the weighted graph of energy landscape.}

\item{r}{SpatRaster object, the energy landscape calculated using the
enerscape package.}

\item{locations}{SpatVector, locations.}
}
\value{
list with:

  - 'costs', matrix of the total costs of travel along the least-cost paths.
  - 'lcps', the SpatVector (lines) of the least-cost paths.
}
\description{
Calculate least-cost paths
}
\details{
The parameters 'locations' must be a SpatVector with points
  geometry. These points are the locations between which least-cost paths
  will be calculated pairwise.
}
